+++ js/ui/ui-init.js
@@ -1,7 +1,4 @@
-import { setupColorUI } from './ui-color.js';
-import { setupSetUI } from './ui-set.js';
 import { setupResetUI } from './ui-reset.js';
-import { setupSubmenuUI } from './submenu/index.js';
 import { setupExportUI } from './ui-export.js';
 import { setupControlsUI } from './ui-controls.js';
 import { setupRoomUI } from './ui-room.js';
@@ -13,10 +10,7 @@ import { setupPresetsUI } from './ui-presets.js';
 
 export function setupUI() {
   try {
-    setupColorUI();
-    setupSetUI();
     setupResetUI();
-    setupSubmenuUI();
     setupExportUI();
     setupControlsUI();
     setupRoomUI();